refactor(web): move PersonaUpsertParameters to svc.ts, types.ts is domain-only
